+++ backuppc-nomad/backuppc-nomad.d/local/share/cook/bin/configure-backuppc.sh
@@ -9,7 +9,7 @@ set -e
 # shellcheck disable=SC3040
 set -o pipefail
 
-export PATH=/usr/local/bin:$PATH
+export PATH="/usr/local/bin:$PATH"
 
 SCRIPT=$(readlink -f "$0")
 TEMPLATEPATH=$(dirname "$SCRIPT")/../templates
@@ -49,4 +49,3 @@ pwd_mkdb -p /etc/master.passwd
 
 # enable backuppc
 service backuppc enable
-