+++ haproxy-minio/haproxy-minio.d/local/share/cook/bin/configure-varnish-exporter.sh
@@ -10,7 +10,7 @@ set -e
 set -o pipefail
 
 # shellcheck disable=SC2086
-export PATH=/usr/local/bin:$PATH
+export PATH="/usr/local/bin:$PATH"
 
 # no steps needed really