+++ haproxy-sql/haproxy-sql.d/local/share/cook/bin/configure-haproxy-3.sh
@@ -9,7 +9,7 @@ set -e
 # shellcheck disable=SC3040
 set -o pipefail
 
-export PATH=/usr/local/bin:$PATH
+export PATH="/usr/local/bin:$PATH"
 
 # make directories
 mkdir -p /usr/local/etc/haproxy