+++ hugo-nginx/hugo-nginx.d/local/share/cook/bin/configure-cron.sh
@@ -9,7 +9,7 @@ set -e
 # shellcheck disable=SC3040
 set -o pipefail
 
-export PATH=/usr/local/bin:$PATH
+export PATH="/usr/local/bin:$PATH"
 
 # setup cronjob for once a day 3am
 echo "0 3 * * *   root   /root/bin/mirrorsync.sh" >> /etc/crontab