+++ nextcloud-nginx-nomad/nextcloud-nginx-nomad.d/local/share/cook/bin/configure-cron.sh
@@ -9,7 +9,7 @@ set -e
 # shellcheck disable=SC3040
 set -o pipefail
 
-export PATH=/usr/local/bin:$PATH
+export PATH="/usr/local/bin:$PATH"
 
 # setup nextcloud cronjob
 echo "*/15 * * * *  www  /usr/local/bin/php -f /usr/local/www/nextcloud/cron.php" >> /etc/crontab