+++ nginx-rsync-ssh/nginx-rsync-ssh.d/local/share/cook/bin/run-postsetup.sh
@@ -9,7 +9,7 @@ set -e
 # shellcheck disable=SC3040
 set -o pipefail
 
-export PATH=/usr/local/bin:$PATH
+export PATH="/usr/local/bin:$PATH"
 
 # if post setup script copied in, make executable and run
 if [ -f /root/postsetup.sh ]; then