+++ traefik-consul/traefik-consul.d/local/share/cook/bin/configure-openssl.sh
@@ -9,7 +9,7 @@ set -e
 # shellcheck disable=SC3040
 set -o pipefail
 
-export PATH=/usr/local/bin:$PATH
+export PATH="/usr/local/bin:$PATH"
 
 # create local ssl directory
 mkdir -p /usr/local/etc/ssl/