+++ traumadrill/traumadrill.d/local/share/cook/bin/configure-local-unbound.sh
@@ -7,7 +7,7 @@ set -e
 # shellcheck disable=SC3040
 set -o pipefail
 
-export PATH=/usr/local/bin:$PATH
+export PATH="/usr/local/bin:$PATH"
 
 echo 'server:
   do-not-query-localhost: no