+++ wordpress-nginx-nomad/wordpress-nginx-nomad.d/local/share/cook/bin/install-wordpress.sh
@@ -9,7 +9,7 @@ set -e
 # shellcheck disable=SC3040
 set -o pipefail
 
-export PATH=/usr/local/bin:$PATH
+export PATH="/usr/local/bin:$PATH"
 
 # install wordpress
 pkg install -y wordpress || true