+++ mailhub-potluck/CHANGELOG.md
@@ -4,6 +4,8 @@
 * Update to 2025Q1 for pkg sources
 * Update for rsync security issue
 * Fix sa-training script for path
+* Add script to requeue the 'on hold' mail
+* Fix issue with shellcheck for PATH
 
 ---