+++ mailhub-potluck/CHANGELOG.md
@@ -1,3 +1,10 @@
+0.21
+
+* Version bump for new base image
+* Add certificate expiry notices to alertmanager host
+
+---
+
 0.20
 
 * Version bump for new base image